Some changes

+++ resources.py
@@ -1,6 +1,6 @@
 from flask_restful import Resource, reqparse
 from models import UserModel, RevokedTokenModel, ContentModel
-from flask_jwt_extended import (create_access_token, create_refresh_token, jwt_required, jwt_refresh_token_required, get_jwt_identity, get_raw_jwt, get_current_user)
+from flask_jwt_extended import (create_access_token, create_refresh_token, jwt_required, jwt_refresh_token_required, get_jwt_identity, get_raw_jwt, current_user)
 
 parser = reqparse.RequestParser()
 parser.add_argument('username', help = 'This field cannot be blank', required = False)
@@ -39,8 +39,8 @@ class UserLogin(Resource):
             return {'message' : 'User {} doesn\'t exist'.format(data['username'])}
         
         if UserModel.verify_hash(data['password'], current_user.password, ):
-            access_token = create_access_token(data['username'])
-            refresh_token = create_refresh_token(identity = data['username'])
+            access_token = create_access_token(identity = current_user.id)
+            refresh_token = create_refresh_token(identity = current_user.id)
             return {
                 'message' : 'Logged in as {}'.format(current_user.username),
                 'access_token' : access_token,
@@ -92,19 +92,25 @@ class OneUser(Resource):
     @jwt_required
     def put(self, user_id):
         data = parser.parse_args()
-        current_user = get_current_user()
-        return {'id' : current_user['id']}
-        #UserModel.edit_user(data, user_id)
+        current_user = get_jwt_identity()
+        if current_user == int(user_id):
+            return UserModel.edit_user(data, user_id) 
+        else:
+            return {'message' : 'Something went wrong!'}
     
     @jwt_required
     def delete(self, user_id):
-        return UserModel.delete_one(user_id)
+        current_user = get_jwt_identity()
+        if current_user == int(user_id):
+            return UserModel.delete_one(user_id)
+        else:
+            return {'message' : 'Something went wrong!'}
+
 
 class AllContents(Resource):
     def get(self):
         return ContentModel.return_all()
     
-    @jwt_required
     def delete(self):
         return ContentModel.delete_all()
     
@@ -115,20 +121,35 @@ class OneContent(Resource):
     @jwt_required
     def put(self, content_id):
         data = parser.parse_args()
-        return ContentModel.edit_content(data, content_id)
+
+        post = ContentModel.find_by_id(content_id)
+        current_user = get_jwt_identity()
+        
+        if current_user == post.user_id:
+            return ContentModel.edit_content(data, content_id, post)
+        else:
+            return {'message' : 'This is not your data!'}
 
     @jwt_required
     def delete(self, content_id):
-        return ContentModel.delete_one(content_id)
+        post = ContentModel.find_by_id(content_id)
+        current_user = get_jwt_identity()
+        
+        if current_user == post.user_id:
+            return ContentModel.delete_one(content_id, post)
+        else:
+            return {'message' : 'This is not your data!'}
+        
 
 class ContentAdd(Resource):
     @jwt_required
     def post(self):
         data = parser.parse_args()
+        current_user = get_jwt_identity()
 
         new_post = ContentModel(
             text = data['text'],
-            user_id = data['user']
+            user_id = current_user
         )
         try:
             new_post.save_to_db()
